Rename install to setup
It makes it clearer that it's not just about installation, but also
about configuring govuk-docker into a working state.
